+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.auth.CheckLoginInterceptor;
+import roomescape.auth.interceptor.CheckLoginInterceptor;
 import roomescape.auth.resolver.LoginMemberArgumentResolver;
 import roomescape.auth.service.AuthService;
